+++ src/zfit_physics/models/pdf_cruijff.py
@@ -73,26 +73,38 @@ class Cruijff(zfit.pdf.BasePDF):
             sigmar: Right width parameter.
             alphar: Right tail acceleration parameter.
         obs: |@doc:pdf.init.obs| Observables of the
-           model. This will be used as the default space of the PDF and,
-           if not given explicitly, as the normalization range.
+               model. This will be used as the default space of the PDF and,
+               if not given explicitly, as the normalization range.
 
-           The default space is used for example in the sample method: if no
-           sampling limits are given, the default space is used.
+               The default space is used for example in the sample method: if no
+               sampling limits are given, the default space is used.
 
-           The observables are not equal to the domain as it does not restrict or
-           truncate the model outside this range. |@docend:pdf.init.obs|
+               If the observables are binned and the model is unbinned, the
+               model will be a binned model, by wrapping the model in a
+               :py:class:`~zfit.pdf.BinnedFromUnbinnedPDF`, equivalent to
+               calling :py:meth:`~zfit.pdf.BasePDF.to_binned`.
+
+               If the observables are binned and the model is unbinned, the
+               model will be a binned model, by wrapping the model in a
+               :py:class:`~zfit.pdf.BinnedFromUnbinnedPDF`, equivalent to
+               calling :py:meth:`~zfit.pdf.BasePDF.to_binned`.
+
+               The observables are not equal to the domain as it does not restrict or
+               truncate the model outside this range. |@docend:pdf.init.obs|
         extended: |@doc:pdf.init.extended| The overall yield of the PDF.
-           If this is parameter-like, it will be used as the yield,
-           the expected number of events, and the PDF will be extended.
-           An extended PDF has additional functionality, such as the
-           ``ext_*`` methods and the ``counts`` (for binned PDFs). |@docend:pdf.init.extended|
+               If this is parameter-like, it will be used as the yield,
+               the expected number of events, and the PDF will be extended.
+               An extended PDF has additional functionality, such as the
+               ``ext_*`` methods and the ``counts`` (for binned PDFs). |@docend:pdf.init.extended|
         norm: |@doc:pdf.init.norm| Normalization of the PDF.
-           By default, this is the same as the default space of the PDF. |@docend:pdf.init.norm|
-        name: |@doc:pdf.init.name| Human-readable name
-           or label of
-           the PDF for better identification.
-           Has no programmatical functional purpose as identification. |@docend:pdf.init.name|
-        label: |@doc:pdf.init.label| A human readable label to identify the PDF. |@docend:pdf.init.label|
+               By default, this is the same as the default space of the PDF. |@docend:pdf.init.norm|
+        name: |@doc:pdf.init.name| Name of the PDF.
+               Maybe has implications on the serialization and deserialization of the PDF.
+               For a human-readable name, use the label. |@docend:pdf.init.name|
+        label: |@doc:pdf.init.label| Human-readable name
+               or label of
+               the PDF for a better description, to be used with plots etc.
+               Has no programmatical functional purpose as identification. |@docend:pdf.init.label|
         """
         params = {"mu": mu, "sigmal": sigmal, "alphal": alphal, "sigmar": sigmar, "alphar": alphar}
         super().__init__(obs=obs, params=params, extended=extended, norm=norm, name=name, label=label)