+++ src/zfit_physics/models/pdf_novosibirsk.py
@@ -149,6 +149,16 @@ class Novosibirsk(zfit.pdf.BasePDF):
                The default space is used for example in the sample method: if no
                sampling limits are given, the default space is used.
 
+               If the observables are binned and the model is unbinned, the
+               model will be a binned model, by wrapping the model in a
+               :py:class:`~zfit.pdf.BinnedFromUnbinnedPDF`, equivalent to
+               calling :py:meth:`~zfit.pdf.BasePDF.to_binned`.
+
+               If the observables are binned and the model is unbinned, the
+               model will be a binned model, by wrapping the model in a
+               :py:class:`~zfit.pdf.BinnedFromUnbinnedPDF`, equivalent to
+               calling :py:meth:`~zfit.pdf.BasePDF.to_binned`.
+
                The observables are not equal to the domain as it does not restrict or
                truncate the model outside this range. |@docend:pdf.init.obs|
             extended: |@doc:pdf.init.extended| The overall yield of the PDF.
@@ -162,9 +172,9 @@ class Novosibirsk(zfit.pdf.BasePDF):
                Maybe has implications on the serialization and deserialization of the PDF.
                For a human-readable name, use the label. |@docend:pdf.init.name|
             label: |@doc:pdf.init.label| Human-readable name
-                or label of
-                the PDF for a better description, to be used with plots etc.
-                Has no programmatical functional purpose as identification. |@docend:pdf.init.label|
+               or label of
+               the PDF for a better description, to be used with plots etc.
+               Has no programmatical functional purpose as identification. |@docend:pdf.init.label|
         """
         params = {"mu": mu, "sigma": sigma, "lambd": lambd}
         super().__init__(obs=obs, params=params, name=name, extended=extended, norm=norm, label=label)