+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Develop
 
 Major Features and Improvements
 -------------------------------
+- `TF-PWA <https://tf-pwa.readthedocs.io/en/latest/>`_ support for loss functions. Minimizer can directly minimize the loss function of a model.
 
 Breaking changes
 ------------------